Support (hacks) for latest Mojo::Log improvements (sort of)
This module replaces Mojo::Log instead of using roles to alter bahavior,
which is a better design choice. Mojo::Log changed it's interface a bit,
and Mojolicious expects that, and so things like Test::Mojo fail.

The better solution to all this is to reimplement this module using
Mojo::Log::Role::AttachLogger, but that would be a fairly significant
rewrite to this module because this module does a lot of things and
stuff, and honestly, I'm lazy.

So I'm hacking up this module a bit more to support the latest
Mojo::Log. This is the fastest solution for now. I may come back to this
later to do a longer-term-safer solution.